✨ feat :  헤더 미니헤더 통합 및 위치 고정
