+++ src/components/layout/Layout.tsx
@@ -1,7 +1,5 @@
 import { useRouter } from "next/router";
-import { useEffect, useState } from "react";
 import colors from "../../../styles";
-import { GetScreenWidth } from "../../assets/getScreenWidth";
 import TabBar from "./TabBar";
 
 interface LayoutProps {
@@ -9,7 +7,7 @@ interface LayoutProps {
 }
 
 export default function Layout({ children }: LayoutProps) {
-  const { pathname, route } = useRouter();
+  const { pathname } = useRouter();
   const viewType =
     pathname.includes("/add") ||
     pathname.includes("/detail") ||
@@ -20,8 +18,6 @@ export default function Layout({ children }: LayoutProps) {
     pathname.includes("/onboarding")
       ? "sub"
       : "main";
-
-  const screenWidth = GetScreenWidth();
 
   return (
     <>
@@ -35,19 +31,13 @@ export default function Layout({ children }: LayoutProps) {
         .container {
           max-width: 390px;
           width: 100vw;
-          // display: flex;
-          // flex-direction: column;
-          // justify-content: center;
-          // align-items: center;
-          // padding-left: 32px;
           background: ${colors.grayWhite};
-          // background: ${colors.grayBackgroundSub};
-          // width: ${screenWidth}px;
         }
         .content {
           padding: 0px 20px;
         }
         .main {
+          padding-top: 110px;
           padding-bottom: 60px;
         }
         .sub {
